backend/requirements.txt
# FastAPI
fastapi==0.116.1
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Image Processing
opencv-python==4.11.0.86
Pillow==11.0.0

# OCR
paddlepaddle==3.0.0b2
paddleocr==2.10.0

# LLM SDKs
google-genai>=0.3.0

# Data Processing
pydantic==2.11.9
pydantic-settings==2.10.1
python-dateutil==2.9.0
dateparser==1.2.0

# Utilities
python-dotenv==1.0.1
